requirements.txt
matplotlib==3.5.0
numpy==1.21.4
pandas==1.3.4
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
statsmodels==0.13.1
tqdm==4.62.3
umap-learn==0.5.2
